+++ config/dev.exs
@@ -1,0 +1,5 @@
+import Config
+
+config :logger,
+  backends: [:console],
+  compile_time_purge_matching: [[level_lower_than: :debug]]