+++ test/lib/crawler/fetcher_test.exs
@@ -55,7 +55,7 @@ defmodule Crawler.FetcherTest do
       |> Map.merge(%{url: url})
       |> Fetcher.fetch()
 
-    assert fetcher == {:error, "Failed to fetch #{url}, status code: 500"}
+    assert fetcher == {:warn, "Failed to fetch #{url}, status code: 500"}
     refute Store.find(url).body
   end
 
@@ -73,7 +73,7 @@ defmodule Crawler.FetcherTest do
       |> Map.merge(%{url: url, timeout: 50})
       |> Fetcher.fetch()
 
-    assert fetcher == {:error, "Failed to fetch #{url}, reason: :timeout"}
+    assert fetcher == {:warn, "Failed to fetch #{url}, reason: :timeout"}
     refute Store.find(url).body
   end
 
@@ -90,7 +90,7 @@ defmodule Crawler.FetcherTest do
         |> Map.merge(%{url: url, timeout: 100, retrier: Retrier})
         |> Fetcher.fetch()
 
-      assert fetcher == {:error, "Failed to fetch #{url}, status code: 500"}
+      assert fetcher == {:warn, "Failed to fetch #{url}, status code: 500"}
       refute Store.find(url).body
     end)
   end